Use bf16

+++ blackwell/39-fp8-mma-2cta.py
@@ -19,7 +19,7 @@ assert N % 256 == 0
 print('Generating inputs...')
 A = (torch.randn(M, K, dtype=torch.bfloat16, device="cuda:0") / K ** 0.25).to(torch.float8_e4m3fn)
 B = (torch.randn(N, K, dtype=torch.bfloat16, device="cuda:0") / K ** 0.25).to(torch.float8_e4m3fn)
-C = torch.zeros(M, N, dtype=torch.float16, device="cuda:0")
+C = torch.zeros(M, N, dtype=torch.bfloat16, device="cuda:0")
 
 # Run kernel
 print('Launching kernel...')
@@ -27,7 +27,7 @@ kernel(A, B, C)
 torch.cuda.synchronize()
 
 # Check correctness
-C_ref = torch.matmul(A.to(torch.float16), B.T.to(torch.float16))
+C_ref = torch.matmul(A.to(torch.bfloat16), B.T.to(torch.bfloat16))
 assert C_ref.dtype == C.dtype
 abs_diff = torch.abs(C_ref - C)
 print(f"Max absolute difference: {abs_diff.max()}")
